refactor: conform to standards, remove dead code
