+++ ui/controls/view.go
@@ -5,17 +5,6 @@ import (
 	"strings"
 )
 
-var roundedBorder = func() lipgloss.Style {
-	return lipgloss.NewStyle().
-		BorderStyle(lipgloss.RoundedBorder())
-}()
-var controlStyle = func() lipgloss.Style {
-	b := lipgloss.RoundedBorder()
-	b.Left = "┤"
-	b.Right = "├"
-	return roundedBorder.BorderStyle(b)
-}()
-
 // View renders the model's content if it is visible, aligning it horizontally and ensuring it fits within the specified width.
 func (m Model) View() string {
 	if !m.IsVisible {
