+++ ui/pages/accounts/style.go
@@ -1,0 +1,5 @@
+package accounts
+
+import "github.com/charmbracelet/lipgloss"
+
+var green = lipgloss.NewStyle().Foreground(lipgloss.Color("10"))
