+++ ui/pages/keys/cmds.go
@@ -1,0 +1,21 @@
+package keys
+
+import (
+	"github.com/algorandfoundation/hack-tui/api"
+	tea "github.com/charmbracelet/bubbletea"
+)
+
+type DeleteKey *api.ParticipationKey
+
+func EmitDeleteKey(key *api.ParticipationKey) tea.Cmd {
+	return func() tea.Msg {
+		return DeleteKey(key)
+	}
+}
+
+// EmitKeySelected waits for and retrieves a new set of table rows from a given channel.
+func EmitKeySelected(key *api.ParticipationKey) tea.Cmd {
+	return func() tea.Msg {
+		return key
+	}
+}
