+++ ui/pages/keys/style.go
@@ -1,0 +1,5 @@
+package keys
+
+import "github.com/charmbracelet/lipgloss"
+
+var green = lipgloss.NewStyle().Foreground(lipgloss.Color("10"))
