--- ui/utils.go
@@ -1,28 +1,0 @@
-package ui
-
-import (
-	"github.com/charmbracelet/bubbles/table"
-	tea "github.com/charmbracelet/bubbletea"
-)
-
-// waitForUint64 handles an uint64 subscription channel as a tea.Message
-func waitForUint64(sub chan uint64) tea.Cmd {
-	return func() tea.Msg {
-		return <-sub
-	}
-}
-
-// waitForTableRows handles an uint64 subscription channel as a tea.Message
-func waitForTableRows(sub chan []table.Row) tea.Cmd {
-	return func() tea.Msg {
-		return <-sub
-	}
-}
-
-// hidden returns 0 when the width is greater than the fill
-func hidden(width int, fillSize int) int {
-	if fillSize < width {
-		return 0
-	}
-	return width
-}
